More files for utils' tests
