+++ UtilsTests/UtilsTests.vcxproj.filters
@@ -72,9 +72,6 @@
     <ClCompile Include="..\utils\include\WindowsUtils.cpp">
       <Filter>Исходные файлы</Filter>
     </ClCompile>
-    <ClCompile Include="src\Tests.cpp">
-      <Filter>Исходные файлы</Filter>
-    </ClCompile>
     <ClCompile Include="..\utils\include\Utils.cpp">
       <Filter>Исходные файлы</Filter>
     </ClCompile>
@@ -82,6 +79,18 @@
       <Filter>Исходные файлы</Filter>
     </ClCompile>
     <ClCompile Include="..\utils\include\CEvent.cpp">
+      <Filter>Исходные файлы</Filter>
+    </ClCompile>
+    <ClCompile Include="src\CXMLDataReaderTest.cpp">
+      <Filter>Исходные файлы</Filter>
+    </ClCompile>
+    <ClCompile Include="src\CStringCleanerTest.cpp">
+      <Filter>Исходные файлы</Filter>
+    </ClCompile>
+    <ClCompile Include="src\CNumericTypesParserTest.cpp">
+      <Filter>Исходные файлы</Filter>
+    </ClCompile>
+    <ClCompile Include="src\CStrToNumWrapperTest.cpp">
       <Filter>Исходные файлы</Filter>
     </ClCompile>
   </ItemGroup>
